애니메이션 처리

+++ src/player.hpp
@@ -27,12 +27,9 @@ private:
   void collision(const std::string& direction);
   void move(int speed);
 
-// private:
-//   py::Rect hitbox;
-
 private:
   std::string status;
-  int frame_index;
+  float frame_index;
   float animation_speed;
   py::Vector2i direction;
   int speed;
